Add files via upload
CPCS425: 
The objective of this course is to provide an introduction to
information security in computer networks, with a focus on
providing basic knowledge of the technical and operational
issues of modern cryptosystems and their related standards. 

CPCS463:
The objective of this course is to introduce the fundamentals of
Computer Security Systems and the potential risks and
vulnerabilities in these therein.